% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_files.R
\name{all_files}
\alias{all_files}
\title{Returns the names of the photos files}
\usage{
all_files(from = "~/Pictures/Photos/", relative = TRUE)
}
\arguments{
\item{from}{A path (character string).}

\item{relative}{Boolean value that specifies whether the returned names of
the files are relative to 'from'. If FALSE, then the names will contain
'from'.}
}
\value{
Returns a vector of character strings, each entry being the name of a
file.
}
\description{
Returns the names of the photos files
}
\examples{
# all_files()
}
