% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_types.R
\name{extract_types}
\alias{extract_types}
\title{Filter the files of a given extension}
\usage{
extract_types(files, ext = "ORF")
}
\arguments{
\item{files}{A vector of character strings that contains the names of the
files (with extensions).}

\item{ext}{A character string specifying the extension to look for.}
}
\value{
Returns the subset of 'files' with the extension 'ext'.
}
\description{
Filter the files of a given extension
}
\examples{
# the names of all the files:
# files <- all_files()
# the names of the ORF files:
# extract_types(files, "ORF")
# the names of the JPG files:
# extract_types(files, "JPG")
}
