% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_types.R
\name{files_types}
\alias{files_types}
\title{Returns the types of photo files}
\usage{
files_types(files)
}
\arguments{
\item{files}{A vector of character strings that contains the names of the
files (with extensions).}
}
\value{
A table with the number of files for each file type, identified from
files extensions.
}
\description{
Returns the types of photo files
}
\examples{
# files_types(all_files())
}
