% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpg_from_camera.R
\name{jpg_from_camera}
\alias{jpg_from_camera}
\title{Returns JPG files generated by the camera}
\usage{
jpg_from_camera(from = "~/Pictures/Photos/", raw_ext = "ORF", jpg_ext = "JPG")
}
\arguments{
\item{from}{A path (character string).}

\item{raw_ext}{A character string specifying the extension used for RAW files.}

\item{jpg_ext}{A character string specifying the extension used for JPG files.}
}
\value{
Returns a vector of character strings with the full names of the JPG
files that were generated by the camera.
}
\description{
Returns JPG files generated by the camera
}
\examples{
# jpg_from_camera()
}
