% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidecar_files.R
\name{sidecar_files}
\alias{sidecar_files}
\title{Retrieves the sidecar files of a particular type}
\usage{
sidecar_files(files, sidecar_ext = ".dop")
}
\arguments{
\item{files}{A vector of character strings that contains the names of the
files (with extensions).}

\item{sidecar_ext}{A character string specifying the extension, with the dot.}
}
\value{
Returns the names of the sidecar files.
}
\description{
Retrieves the sidecar files of a particular type
}
\examples{
# jpg_with_orf <- jpg_from_camera()
# dop_files <- sidecar_files(jpg_with_orf, ".dop")
# to_remove <- c(jpg_with_orf, dop_files)
# # file.remove(to_remove)
}
